+++ includes/providers/class-frontend-service-provider.php
@@ -28,8 +28,16 @@ class FrontendServiceProvider extends AbstractServiceProvider {
 	 * Boot the service provider.
 	 */
 	public function boot(): void {
-		if ( ! is_admin() ) {
-			$this->container->get( 'frontend' );
+		parent::boot();
+
+		if ( ! is_admin() || wp_doing_ajax() ) {
+			// Register hook to initialize after WordPress is loaded
+			add_action(
+				'init',
+				function () {
+					$this->container->get( 'frontend' )->init();
+				}
+			);
 		}
 	}
 }